replace filon #6

+++ src/cxx/Trj/libinc/Filon.hh
@@ -22,21 +22,16 @@ void sin_integral_single(unsigned x_panels,double *xVec,double *yVec,
 void cos_integral_single(unsigned x_panels,double *xVec,double *yVec,
                         double time, double &cos_single);
 
-void cal_integral(unsigned massNum, double temperature, unsigned x_panels,double *xVec,double *yVec,
+void cal_integral(double massNum, double temperature, unsigned x_panels,double *xVec,double *yVec,
                       unsigned t_length, double *timeVec, double *gamma_classic,
                       double *gamma_quantum_real, double *gamma_quantum_imag);
 
-void cal_limit(unsigned massNum, double temperature, double *xVec, double *yVec,
+void cal_limit(double massNum, double temperature, double *xVec, double *yVec,
                       unsigned t_length, double *timeVec, double *limit_value,
                       double *limit_value_real, double *limit_value_imag);
 
 void cal_PDF(double rho, unsigned x_panels,double *xVec,double *yVec,
                           unsigned r_length, double *rVec, double* pdfVec);
-//void fftconvolveCXX(double *a1, double *a2,
-//			double *y, double dt);
-
-//void fftconvolveCXX(const std::vector<double>& a1, const std::vector<double>& a2,
-//                      std::vector<double>& y, double dt);
 
 #ifdef __cplusplus
 }
